+++ packages/base-monaco-editor/src/index.tsx
@@ -11,6 +11,7 @@ import {
 } from './helper';
 
 export * from './monaco';
+export * from './controller';
 
 const SingleMonacoEditor = (props: ISingleMonacoEditorProps) => {
   const { onChange, enableOutline, width, height, language, supportFullScreen } = props;
@@ -29,7 +30,7 @@ const SingleMonacoEditor = (props: ISingleMonacoEditorProps) => {
   } = useEditor<editor.IStandaloneCodeEditor>('single', props);
   const subscriptionRef = useRef(null);
 
-  const className = classNames('ve-code-control', props.className, {
+  const className = classNames('lc-code-control', props.className, {
     've-focused': focused,
     've-outline': enableOutline,
   });
@@ -139,7 +140,7 @@ const DiffMonacoEditor = (props: IDiffMonacoEditorProps) => {
     props,
   );
 
-  const className = classNames('ve-code-control', props.className, {
+  const className = classNames('lc-code-control', props.className, {
     've-focused': focused,
     've-outline': enableOutline,
   });