+++ packages/plugin-code-editor/.eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+  extends: [
+    'eslint-config-ali/typescript/react',
+    'prettier',
+    'prettier/@typescript-eslint',
+    'prettier/react',
+  ],
+  rules: {
+    'no-param-reassign': 0,
+  },
+};