+++ packages/plugin-code-editor/src/utils/transform.ts
@@ -1,4 +1,3 @@
-
 import { transform } from './babel';
 import { TransformResult } from '../types';