+++ packages/plugin-datasource-pane/src/pane/DataSourcePane.tsx
@@ -9,8 +9,6 @@ import {
 } from '@alilc/lowcode-datasource-types';
 import { Button, Message, Dialog } from '@alifd/next';
 import _isArray from 'lodash/isArray';
-import { HTML5Backend } from 'react-dnd-html5-backend';
-import { DndProvider } from 'react-dnd';
 import { createStateMachine } from '../utils/stateMachine';
 import { DataSourcePaneContext } from '../utils/panel-context';
 import { DataSourceFilter } from '../components/DataSourceFilter';
@@ -72,11 +70,11 @@ export class DataSourcePane extends PureComponent<
     listMode: DataSourcePanelMode.NORMAL,
   };
 
-  detailRef? = createRef<DataSourceForm>();
+  detailRef ?= createRef<DataSourceForm>();
 
-  exportRef? = createRef<ExportDetail>();
+  exportRef ?= createRef<ExportDetail>();
 
-  importRef? = createRef<ImportDetail>();
+  importRef ?= createRef<ImportDetail>();
 
   private send = (...args: any[]) => {
     this.context.stateService.send(...args);
@@ -85,7 +83,8 @@ export class DataSourcePane extends PureComponent<
   componentDidMount() {
     this.serviceS = this.context?.stateService?.subscribe?.((state: any) => {
       this.setState({ current: state });
-      if (state.changed && state.value === 'idle') {
+      // 监听导入成功事件
+      if (state.changed && (state.value === 'idle' || state.event?.type === 'FINISH_IMPORT')) {
         // TODO add hook
         this.props.onSchemaChange?.({
           list: state.context.dataSourceList,
@@ -221,9 +220,9 @@ export class DataSourcePane extends PureComponent<
           return;
         }
         const repeatedDataSourceList = data.filter(
-          (item) => !!this.state.current.dataSourceList.find(
-              (dataSource: DataSourceConfig) => dataSource.id === item.id,
-            ),
+          (item) => !!this.state.current.context.dataSourceList.find(
+            (dataSource: DataSourceConfig) => dataSource.id === item.id,
+          ),
         );
         if (repeatedDataSourceList.length > 0) {
           Dialog.confirm({
@@ -237,13 +236,15 @@ export class DataSourcePane extends PureComponent<
           return;
         }
         importDataSourceList();
+      }).catch(err => {
+        console.warn(err?.message);
       });
     }
   };
 
   renderDetail = () => {
     const { current } = this.state;
-    const { dataSourceTypes = [] } = this.props;
+    const { dataSourceTypes = [], importPlugins = [], exportPlugins = [], formComponents = [] } = this.props;
     let content = null;
 
     if (current.matches('detail.edit')) {
@@ -279,7 +280,11 @@ export class DataSourcePane extends PureComponent<
     } else if (current.matches('detail.import')) {
       // TODO
       // pluginName
-      content = <ImportDetail dataSourceTypes={dataSourceTypes} ref={this.importRef} />;
+      const currentPluginName = current.context.detail?.data?.pluginName?.name;
+      const importPlugin = importPlugins?.find((item) => item.name === currentPluginName);
+      const Component = (importPlugin ? importPlugin.component : ImportDetail) as React.ElementType;
+      content = <Component dataSourceTypes={dataSourceTypes} ref={this.importRef} />;
+
     } else if (current.matches('detail.export')) {
       // TODO
       content = (
@@ -322,7 +327,7 @@ export class DataSourcePane extends PureComponent<
       className = '',
       helpLink = '',
       dataSourceTypes = [],
-      importPlugins,
+      importPlugins = [],
     } = this.props;
     const { current, listMode } = this.state;
 
@@ -343,54 +348,52 @@ export class DataSourcePane extends PureComponent<
     const isEmpty = current.context.dataSourceList.length === 0;
 
     return (
-      <DndProvider backend={HTML5Backend}>
-        <div
-          className={cn(generateClassName('container'), className)}
-          style={style}
-        >
-          <div className={generateClassName('title')}>
-            数据源
-            {helpLink && <Button component="a" href={helpLink}>使用帮助</Button>}
-          </div>
-          <DataSourceFilter
-            key={current.context.dataSourceListFilter.renderKey}
-            dataSourceTypes={dataSourceTypes}
-            onFilter={this.handleFilterChange}
-          />
-          <DataSourceOperations
-            dataSource={current.context.dataSourceList}
-            empty={isEmpty}
-            importPlugins={importPlugins}
-            dataSourceTypes={dataSourceTypes}
-            onCreate={this.handleCreate}
-            onStartExport={this.handleStartExport}
-            onCancelExport={this.handleCancelExport}
-            onFinishExport={this.handleFinishExport}
-            onStartSort={this.handleStartSort}
-            onCancelSort={this.handleCancelSort}
-            onFinishSort={this.handleFinishSort}
-            onImport={this.handleImport}
-            mode={mode}
-            selectedList={current.context.export.selectedDataSourceIdList}
-          />
-          <div className={generateClassName('list')}>
-            <DataSourceList
-              renderItemInfoTags={renderDataSourceInfoTags}
-              dataSource={current.context.dataSourceList.filter((i: DataSourceConfig) => {
-                return (
-                  i.id.indexOf(current.context.dataSourceListFilter.keyword) !==
-                    -1 &&
-                  (!current.context.dataSourceListFilter.dataSourceType ||
-                    current.context.dataSourceListFilter.dataSourceType ===
-                      i.type)
-                );
-              })}
-              onOperationClick={this.handleOperationClick}
-            />
-          </div>
-          {this.renderDetail()}
+      <div
+        className={cn(generateClassName('container'), className)}
+        style={style}
+      >
+        <div className={generateClassName('title')}>
+          数据源
+          {helpLink && <Button component="a" href={helpLink}>使用帮助</Button>}
         </div>
-      </DndProvider>
+        <DataSourceFilter
+          key={current.context.dataSourceListFilter.renderKey}
+          dataSourceTypes={dataSourceTypes}
+          onFilter={this.handleFilterChange}
+        />
+        <DataSourceOperations
+          dataSource={current.context.dataSourceList}
+          empty={isEmpty}
+          importPlugins={importPlugins}
+          dataSourceTypes={dataSourceTypes}
+          onCreate={this.handleCreate}
+          onStartExport={this.handleStartExport}
+          onCancelExport={this.handleCancelExport}
+          onFinishExport={this.handleFinishExport}
+          onStartSort={this.handleStartSort}
+          onCancelSort={this.handleCancelSort}
+          onFinishSort={this.handleFinishSort}
+          onImport={this.handleImport}
+          mode={mode}
+          selectedList={current.context.export.selectedDataSourceIdList}
+        />
+        <div className={generateClassName('list')}>
+          <DataSourceList
+            renderItemInfoTags={renderDataSourceInfoTags}
+            dataSource={current.context.dataSourceList.filter((i: DataSourceConfig) => {
+              return (
+                i.id.indexOf(current.context.dataSourceListFilter.keyword) !==
+                -1 &&
+                (!current.context.dataSourceListFilter.dataSourceType ||
+                  current.context.dataSourceListFilter.dataSourceType ===
+                  i.type)
+              );
+            })}
+            onOperationClick={this.handleOperationClick}
+          />
+        </div>
+        {this.renderDetail()}
+      </div>
     );
   }
 }