+++ packages/plugin-schema/src/index.tsx
@@ -1,15 +1,21 @@
 import * as React from 'react';
-import { ILowCodePluginContext } from '@alilc/lowcode-engine';
+import { IPublicModelPluginContext } from '@alilc/lowcode-types';
 import PluginSchema from './editor';
+import { enUS, zhCN } from './locale';
 
-const plugin = (ctx: ILowCodePluginContext) => {
+const plugin = (ctx: IPublicModelPluginContext, options: any) => {
   return {
-    // 插件名，注册环境下唯一
-    name: 'LowcodePluginAliLowcodePluginSchema',
-    // 依赖的插件（插件名数组）
-    dep: [],
     // 插件的初始化函数，在引擎初始化之后会立刻调用
     init() {
+      const { intl, intlNode, getLocale } = ctx.common.utils.createIntl({
+        'en-US': enUS,
+        'zh-CN': zhCN,
+      });
+      ctx.intl = intl;
+      ctx.intlNode = intlNode;
+      ctx.getLocale = getLocale;
+      const isProjectSchema = (options && options['isProjectSchema']) === true;
+
       // 往引擎增加面板
       ctx.skeleton.add({
         area: 'leftArea',
@@ -25,9 +31,8 @@ const plugin = (ctx: ILowCodePluginContext) => {
         },
         content: () => (
           <PluginSchema
-            project={ctx.project}
-            skeleton={ctx.skeleton}
-            event={ctx.event}
+            pluginContext={ctx}
+            showProjectSchema={isProjectSchema}
           />
         ),
       })
@@ -35,6 +40,19 @@ const plugin = (ctx: ILowCodePluginContext) => {
   };
 };
 
-plugin.pluginName = 'LowcodePluginAliLowcodePluginSchema'
+plugin.pluginName = 'LowcodePluginAliLowcodePluginSchema';
+plugin.meta = {
+  preferenceDeclaration: {
+    title: 'schema 面板配置',
+    properties: [
+      {
+        key: 'isProjectSchema',
+        type: 'boolean',
+        description: '是否是项目级 schema',
+        default: false,
+      },
+    ],
+  },
+};
 
-export default plugin
+export default plugin;